rc-price-tracker/requirements.txt
requests>=2.31.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
apprise>=1.7.0
apscheduler>=3.10.0
selenium>=4.15.0
webdriver-manager>=4.0.0
flask>=3.0.0
gunicorn>=22.0.0
